bowl static

+++ src/main/java/Lesson_6/ForUnitTests.java
@@ -44,7 +44,7 @@ public class ForUnitTests {
         return Math.multiplyExact(a,b);
     }
 
-    public static double division(int a, int b) throws MyArithmeticException {
+    public static int division(int a, int b) throws MyArithmeticException {
         try {
             return a / b;
         } catch (ArithmeticException e) {
